load_tecandata now returns a TecanSet. TecanPlate view_plate() function added

+++ tecanpack/readers.py
@@ -4,6 +4,7 @@ from pathlib import Path
 import importlib.util
 import pandas as pd
 from . import tecanio
+from .tecanio import TecanSet
 
 def determinepath(curpath,cfdict):
     if cfdict['filepath_prefix_type']=='relative':
@@ -19,7 +20,7 @@ def read_load_confile(cfpathstr,refresh_all):
     cfp=Path(cfpathstr)
     cfdir=cfp.parent
     with open(cfp,'r') as f:
-        configgy=yaml.load(f)
+        configgy=yaml.safe_load(f)
     #logdb first:
     logdb_dict=configgy['logdb']
     logdbpath=determinepath(cfdir,logdb_dict)
@@ -41,10 +42,11 @@ def read_load_confile(cfpathstr,refresh_all):
     c=conn.cursor()
 
     if refresh_all:
-        c.execute('''DROP TABLE PLATES''')
-        c.execute('''DROP TABLE LSCRIPTS''')
-        for pklf in os.listdir(pkl_fldrpath):
-            os.remove(os.path.join(pkl_fldrpath,pklf))
+        c.execute('''DROP TABLE IF EXISTS PLATES''')
+        c.execute('''DROP TABLE IF EXISTS LSCRIPTS''')
+        if pkl_fldrpath.exists():
+            for pklf in os.listdir(pkl_fldrpath):
+                os.remove(os.path.join(pkl_fldrpath,pklf))
     c.execute('''CREATE TABLE IF NOT EXISTS PLATES (plateid text, filename text, sheetname text,scriptname text, pklplate glob)''')
     c.execute('''CREATE TABLE IF NOT EXISTS LSCRIPTS (scriptname text, pkl_ctime text, pklfname text)''')
 
@@ -69,6 +71,7 @@ def read_load_confile(cfpathstr,refresh_all):
         except:
             print(f'reloading plates using {lsentry}')
             load_required=True
+#            conn.close()
         if load_required:
             spec=importlib.util.spec_from_file_location(str(ls[0]),ls[0])#'kdfs/dload_dir/load_scripts/testscript.py','kdfs/dload_dir/load_scripts/testscript.py')#
             themodule=importlib.util.module_from_spec(spec)
@@ -76,15 +79,17 @@ def read_load_confile(cfpathstr,refresh_all):
             m2c=getattr(themodule,ls[1])
             plates=m2c()
             for plate in plates:
-                newtuple=(plate.plateid,plate.ifpath.name,plate.expsheet,lsentry,pickle.dumps(plate))
-                c.execute('''INSERT INTO PLATES VALUES (?,?,?,?,?)''',newtuple)
+                newpltuple=(plate.plateid,plate.ifpath.name,plate.expsheet,lsentry,pickle.dumps(plate))
+                c.execute('''INSERT INTO PLATES VALUES (?,?,?,?,?)''',newpltuple)
             pklfname=f'pltscached-{uuid.uuid4().hex}'
             pklpath=pkl_fldrpath / pklfname
+            if pkl_fldrpath.exists()==False:
+                os.mkdir(pkl_fldrpath)
             with open(pklpath,'wb') as f:
                 pickle.dump(plates,f)
             curlstime=os.path.getmtime(pklpath)
             newlstuple=(lsentry,curlstime,pklfname)
-            c.execute('''INSERT INTO LSCRIPTS VALUES (?,?,?)''',newtuple)
+            c.execute('''INSERT INTO LSCRIPTS VALUES (?,?,?)''',newlstuple)
         conn.commit()
         #now read in the plates--
         c.execute('''SELECT * FROM LSCRIPTS WHERE scriptname=(?)''',(lsentry,))
@@ -103,6 +108,8 @@ def load_tecandata(cfpathstr,refresh_all=False):
     Keyword Arguments:
         refresh_all: whether to reset all files for a fresh re-read (default False)
     '''
-    dfs=read_load_confile(cfpath,refresh_all)
-    mergedf=pd.concat([x for x in dfs],ignore_index=True)
-    return mergedf
+    plates=read_load_confile(cfpathstr,refresh_all)
+    tpset=TecanSet(plates)
+    return tpset
+#    mergedf=pd.concat([x for x in dfs],ignore_index=True)
+#    return mergedf
